refactor: extract `buildAgentPool` to `src/agents.ts` to break `config.ts` -> `review.ts` layering inversion
`src/config.ts` (low layer) was importing `buildAgentPool` from `src/review.ts` (orchestration layer) purely to validate the `reviewers` config block against the standard agent pool. Move `AGENT_POOL`, `TRIVIAL_VERIFIER_AGENT`, and `buildAgentPool` to a new minimal `src/agents.ts` module whose only dependency is `ReviewerAgent` from `./types`. `src/config.ts`, `src/index.ts`, `src/review.ts`, and the test files all import these primitives from `./agents` directly so the agent-pool primitives have a single canonical home decoupled from review orchestration. Closes #676.
