+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `openai.ts` no longer defines its own `buildTimeoutDiagnostics`; it now imports the function directly from `cli-utils`, matching how `anthropic.ts` and `gemini.ts` already consume it. The local copy was structurally identical to the exported one, so this is a pure deduplication with no behaviour change ([#722](https://github.com/manki-review/manki/issues/722)).
+- Extracted `AGENT_POOL`, `TRIVIAL_VERIFIER_AGENT`, and `buildAgentPool` from `src/review.ts` into a new low-level `src/agents.ts` module, breaking the layering inversion where `src/config.ts` (low layer) imported from `src/review.ts` (orchestration layer) just to validate `reviewers` names against the agent pool. `src/config.ts`, `src/index.ts`, and the test files now import these primitives directly from `./agents`, and `src/review.ts` consumes them from the same module so the agent-pool primitives have a single canonical home decoupled from review orchestration ([#676](https://github.com/manki-review/manki/issues/676)).
 - Switched to the canonical TypeScript-Action release model: the compiled `dist/` is committed to `main` so both floating (`manki-review/manki@v5`) and immutable exact-version (`manki-review/manki@v5.x.y`) pins resolve to a runnable action. Previously, `release.yml` committed `dist/` onto a one-off SHA and force-updated only the major tag, leaving exact-version tags pointing at source-only commits that failed with `File not found: dist/index.js` on the runner. A new [`check-dist.yml`](.github/workflows/check-dist.yml) workflow runs on every PR and `push` to `main`, rebuilds `dist/`, and fails if the committed copy drifts from a fresh build. `release.yml` is simplified to just force-update the floating major tag and publish the GitHub Release. `dist/**` is added to `.gitattributes` (`linguist-generated=true`, plus `-diff` on the minified entry) so the GitHub PR Files view collapses the bundle diff, and to `.manki.yml` / `codecov.yml` so manki self-reviews and coverage reports skip it. [`CONTRIBUTING.md`](CONTRIBUTING.md) documents the local `npm run build` step contributors must run before opening a PR. Historical `v5.0.0` and `v5.0.1` tags remain source-only and continue to require pinning to `@v5` ([#800](https://github.com/manki-review/manki/issues/800)).
 
 ## [5.1.1] - 2026-05-22
